+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [GH#3](https://github.com/Korbeil/phpstan-generic-rules/pull/3) MbStringRule for string function that have a `mb_*` equivalent
+
+### Changed
+- [GH#4](https://github.com/Korbeil/phpstan-generic-rules/pull/4) DebugRule is faster
 
 ## [0.1.0] - 2019-09-12
 ### Added